refactored io
