+++ nngt/core/ig_graph.py
@@ -30,8 +30,8 @@ import scipy.sparse as ssp
 import nngt
 from nngt.lib import InvalidArgument, nonstring_container, BWEIGHT, is_integer
 from nngt.lib.connect_tools import _cleanup_edges
-from nngt.lib.graph_helpers import _to_np_array, _get_dtype, _get_ig_weights
-from nngt.lib.io_tools import _np_dtype
+from nngt.lib.graph_helpers import _get_dtype, _get_ig_weights
+from nngt.lib.converters import _np_dtype, _to_np_array
 from nngt.lib.logger import _log_message
 from .graph_interface import GraphInterface, BaseProperty
