+++ nngt/core/nngt_graph.py
@@ -31,8 +31,8 @@ from scipy.sparse import coo_matrix, lil_matrix
 import nngt
 from nngt.lib import InvalidArgument, nonstring_container, is_integer
 from nngt.lib.connect_tools import _cleanup_edges
-from nngt.lib.graph_helpers import _get_edge_attr, _to_np_array, _get_dtype
-from nngt.lib.io_tools import _np_dtype
+from nngt.lib.graph_helpers import _get_edge_attr, _get_dtype
+from nngt.lib.converters import _np_dtype, _to_np_array
 from nngt.lib.logger import _log_message
 from .graph_interface import GraphInterface, BaseProperty
