+++ nngt/lib/graph_backends.py
@@ -112,7 +112,7 @@ def use_backend(backend, reloading=True, silent=False):
     if reloading:
         reload(sys.modules["nngt"].analysis.clustering)
         reload(sys.modules["nngt"].analysis.graph_analysis)
-        reload(sys.modules["nngt"].analysis)  # must come after  graph_analysis
+        reload(sys.modules["nngt"].analysis)  # must come after graph_analysis
         reload(sys.modules["nngt"].generation.graph_connectivity)
         reload(sys.modules["nngt"].generation)
